profiles/coreos/base: disable gcc internal bootstrap
Gentoo's GCC build system performs a bootstrap internally, which includes
rebuilding the new GCC with itself, and comparing the result. This steps adds
build-time, and is disabled for cross-compilers (can't execute programs built
with a cross-compiler on the build machine). Since we don't modify gcc in any
way, heavily rely on cross-compilation any way, and have other ways of finding
issues, disable this bootstrap.

We build gcc 3 times in bootstrap_sdk so this should be noticable.